Move useLoadClientScript test cases to separate spec
